requirements.txt
# Core Framework
streamlit==1.31.0
fastapi==0.109.0
uvicorn==0.27.0

# Database
psycopg2-binary==2.9.9
sqlalchemy==2.0.25

# NLP & ML
spacy==3.7.2
en-core-web-sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.7.1/en_core_web_sm-3.7.1-py3-none-any.whl
sentence-transformers==2.3.1
scikit-learn==1.4.0

# PDF/Document Processing
pdfplumber==0.10.3
pdfminer.six==20221105
python-docx==1.1.0
pdf2image==1.17.0
pytesseract==0.3.10
easyocr==1.7.1
Pillow==10.2.0

# Fuzzy Matching & Text Processing
# thefuzz==0.21.0
python-Levenshtein==0.24.0

# Data Processing
pandas==2.2.0
numpy==1.26.3

# Visualization
plotly==5.18.0
matplotlib==3.8.2

# Email Processing
imap-tools==1.5.0

# AI/LLM
groq==0.4.2

# Utilities
python-dotenv==1.0.1
python-multipart==0.0.6

# Logging & Monitoring
colorlog==6.8.2

# HTTP & APIs
requests==2.31.0
httpx==0.26.0

# Date/Time
python-dateutil==2.8.2

# File Handling
pathlib2==2.3.7.post1

# Testing (Optional)
pytest==8.0.0
pytest-cov==4.1.0
